+++ graphics/r_shaders.h
@@ -11,56 +11,61 @@ namespace fresa::graphics
     // · DEFINITIONS ·
     // ···············
 
-    //: shader types
+    //: shader stages
     struct ShaderStageData {
         str_view extension;
         VkShaderStageFlagBits stage;
     };
-
-    enum struct ShaderStages {
+    enum struct ShaderStage {
         VERTEX,
         FRAGMENT,
         COMPUTE
     };
-
     constexpr inline auto shader_stages = std::to_array<ShaderStageData>({
         {"vert", VK_SHADER_STAGE_VERTEX_BIT},
         {"frag", VK_SHADER_STAGE_FRAGMENT_BIT},
         {"comp", VK_SHADER_STAGE_COMPUTE_BIT}
     });
 
+    //: shader descriptor types
+    enum struct ShaderDescriptor {
+        UNIFORM,
+        STORAGE,
+        IMAGE_SAMPLER,
+        INPUT_ATTACHMENT
+    };
+    struct ShaderDescriptorData {
+        str_view name;
+        VkDescriptorType descriptor_type;
+    };
+    constexpr inline auto shader_descriptors = std::to_array<ShaderDescriptorData>({
+        {"uniform", VK_DESCRIPTOR_TYPE_UNIFORM_BUFFER},
+        {"storage", VK_DESCRIPTOR_TYPE_STORAGE_BUFFER},
+        {"image_sampler", VK_DESCRIPTOR_TYPE_COMBINED_IMAGE_SAMPLER},
+        {"input_attachment", VK_DESCRIPTOR_TYPE_INPUT_ATTACHMENT}
+    });
+
     // ·············
     // · DATATYPES ·
     // ·············
 
-    //: shader resource
-    //      a binding for a specific descriptor resource item
-    /*struct ShaderResource {
-        ui16 id;
-        str name;
-        VkDescriptorType type;
-    };*/
-    
-    //: descriptor set
-    //      holds the descriptors and bindings for each glsl set, used for sending data to the shader
-    /*struct DescriptorSet {
+    //: descriptor set layout binding
+    struct DescriptorLayoutBinding {
+        ui32 binding;
         ui32 set;
-        VkDescriptorSetLayout layout;
-        VkDescriptorPool* pool;
-        std::array<VkDescriptorSet, engine_config.vk_frames_in_flight()> descriptors;
-        std::vector<VkDescriptorSetLayoutBinding> bindings;
-        std::vector<ShaderResource> resources;
-    };*/
+        ui32 size;
+        ui32 descriptor_count;
+        VkDescriptorType descriptor_type;
+        ShaderStage stage_flags;
+        str name;
+    };
 
-    //: shader pass
-    //      combination of multiple shader stages to create a pipeline
-    //      holds the list of modules, descriptor sets and the built pipeline
-    /*struct ShaderPass {
-        std::vector<VkShaderModule> stages;
-        std::vector<DescriptorSet> descriptors;
-        VkPipelineLayout pipeline_layout;
-        VkPipeline pipeline;
-    };*/
+    //: shader module
+    struct ShaderModule {
+        VkShaderModule module;
+        ShaderStage stage;
+        std::vector<DescriptorLayoutBinding> bindings;
+    };
 
     // ···········
     // · SYSTEMS ·
@@ -68,11 +73,26 @@ namespace fresa::graphics
 
     namespace shader
     {
-        //: read spirv code
-        std::vector<ui32> readSPIRV(str name, ShaderStages stage);
+        //* spirv
 
-        //: create shader module from spirv code
-        VkShaderModule createModule(str name, ShaderStages stage);
+        //: read spirv code from .spv file
+        std::vector<ui32> readSPIRV(str name, ShaderStage stage);
+
+        //: create spirv cross compiler from the code
+        //      this is used to get the reflection data from the shader
+        //      it has to revert all the bits in the shader code since spirv cross takes it that way
+        spv_c::CompilerGLSL createCompiler(std::vector<ui32> code);
+
+        //: use spirv reflection to automatically get the descriptor set layout bindings
+        std::vector<DescriptorLayoutBinding> getDescriptorBindings(const spv_c::CompilerGLSL& compiler, ShaderStage stage);
+
+        //* vulkan shader components
+
+        //: create vulkan shader module from spirv code
+        VkShaderModule createVkShader(const std::vector<ui32>& code);
+
+        //: create shader module object
+        ShaderModule createModule(str name, ShaderStage stage);
 
         //: create a shader pass from one or more shader modules, also does reflection on descriptor layouts
         //- ShaderPass createPass(str name);